# openapi.yaml
openapi: 3.1.0
info:
  title: Zantara Bridge API
  description: |
    Comprehensive API for Google Workspace integration, AI-powered chat, and productivity tools.
    The API provides secure endpoints for Drive, Gmail, Calendar operations, AI analytics, and conversation management.
    
    **Performance Features:**
    - Redis-based caching for Drive operations (5min folder cache, 10min file cache, 3min search cache)
    - Firestore-powered memory system with full-text search and relevance scoring
    - Automatic cache invalidation on data changes
    - Zero-cost optimized architecture with graceful fallbacks
  version: 1.0.0
  contact:
    name: Bali Zero HQ
    email: zero@balizero.com
  license:
    name: Proprietary
servers:
  - url: https://zantara-bridge-v2-prod-1064094238013.asia-southeast2.run.app
    description: Production server (Cloud Run)
  - url: https://zantara-light-bridge-1064094238013.asia-southeast2.run.app
    description: Zantara Light Bridge - Zero-cost optimized (Cloud Run)
  - url: http://localhost:8080
    description: Development server
security:
  - ApiKeyAuth: []
  - UserIdentification: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
      description: API key for authentication (required for most endpoints)
    UserIdentification:
      type: apiKey
      in: header
      name: X-BZ-USER
      description: User identification header (defaults to 'BOSS' if not provided)
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized - Invalid or missing API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        ok:
          type: boolean
          default: false
        error:
          type: string
          description: Error message
        details:
          type: string
          description: Additional error details
      required:
        - error
...
paths:
...
  /call:
    post:
      tags: [AI & Chat]
      summary: Lightweight AI assistant call
      description: |
        Process AI assistant requests with simplified lightweight implementation.
        Supports basic chat, drive actions, and memory operations with cost optimization.
      parameters:
        - name: X-BZ-USER
          in: header
          schema:
            type: string
          description: User identifier for logging (defaults to 'LIGHT_USER')
        - name: X-API-KEY
          in: header
          schema:
            type: string
          description: API key for authentication (optional for this endpoint)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
                  description: User message or content
                  example: "Help me with my task"
                key:
                  type: string
                  description: Action key to determine behavior
                  enum: [chat.simple, drive.upload, memory.save]
                  default: chat.simple
                  example: "chat.simple"
                params:
                  type: object
                  description: Additional parameters for specific actions
                  example: {}
                model:
                  type: string
                  description: AI model to use
                  default: "gpt-4o-mini"
                  example: "gpt-4o-mini"
              required:
                - message
      responses:
        '200':
          description: AI response processed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    default: true
                  service:
                    type: string
                    example: "zantara-light-bridge"
                    description: Service identifier
                  response:
                    type: string
                    description: Processed response text
                    example: "Light bridge processed: Hello from CLI test"
                  timestamp:
                    type: string
                    format: date-time
                    description: Response timestamp
                    example: "2025-09-19T03:19:02.674Z"
        '400':
          description: Bad request - missing message
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    default: false
                  error:
                    type: string
                    example: "Message is required"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /actions/memory/save:
    post:
      tags: [Memory]
      summary: Save memory entry (Firestore Implementation)
      description: |
        Save a memory entry to Firestore with full metadata support.
        Supports tags, categories, relevance scoring, and automatic access tracking.
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                  description: User identifier for memory isolation
                  example: "user-123"
                content:
                  type: string
                  description: Main memory content
                  example: "Meeting notes about quarterly planning"
                title:
                  type: string
                  description: Optional memory title
                  example: "Q4 Planning Meeting"
                tags:
                  type: array
                  items:
                    type: string
                  description: Searchable tags
                  example: ["meeting", "planning", "q4"]
                category:
                  type: string
                  description: Memory category
                  example: "work"
                source:
                  type: string
                  description: Origin of the memory
                  example: "api"
                metadata:
                  type: object
                  properties:
                    tokenCount:
                      type: number
                    contextType:
                      type: string
                      enum: [conversation, document, note, task]
                  description: Additional metadata
              required:
                - userId
                - content
      responses:
        '200':
          description: Memory entry saved successfully to Firestore
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    default: true
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        description: Firestore document ID
                        example: "abc123def456"
                      message:
                        type: string
                        example: "Memory entry saved successfully"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /actions/drive/upload:
    post:
      tags: [Drive]
      summary: Upload file to Drive (with cache invalidation)
      description: |
        Upload a file to Google Drive with impersonation support.
        Automatically invalidates relevant cache entries for folder listings and search results.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                filename:
                  type: string
                content:
                  type: string
                folderId:
                  type: string
                mimeType:
                  type: string
              required:
                - filename
                - content
      responses:
        '200':
          description: File uploaded successfully (cache invalidated)
          headers:
            X-Cache-Status:
              description: Cache invalidation status
              schema:
                type: string
                example: "invalidated"
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  webViewLink:
                    type: string
                  cacheStatus:
                    type: string
                    description: Cache invalidation status
                    example: "folders_invalidated"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...